Get syneccmd back working.
r517 modified GUI_RPC_CLIENT::init to take GUI_RPC_PORT as default value for the port argument. It used to take 0 as default, and inside the function, used GUI_RPC_PORT if the argument was 0.
However, boinccmd code was passing an explicit 0 as port number. I modified boinccmd to pass GUI_RPC_PORT instead of 0, but maybe it's better to revert that part of r517.