Cleanup version.h mess.
- Version information is not present in configure.ac anymore,
other than the required AC_INIT.

- Version info is only in version.h, which is a static file to be edited manually.
It's not generated by the configure script. version.h.in is gone.

- The PACKAGE_* macros defined automatically by autotools
aren't used in the code anymore (there was only boinc_cmd.C using one).
They were also removed from version.h.

- The top source directory is appended to the include path,
since that's where version.h is now.
version.h used to be in the top *build* directory, as it was a generated file.