Move lib unit tests into lib/tests/.
The root tests directory now has only common code to be used from tests,
and the test runner. This code is in a static library so tests can link to it easily.

Each subdirectory with tests will have its own CMakeLists.txt with the tests to run.


Also fixed the automake buildsystem to work with these changes.
Can't wait to get rid of it, though.