Prefs dialog: Change how preferences with text fields embedded in sentences are created.
Before, the code was creating the labels and text fields and adding
them to the dialog directly. This change adds a function that takes
a string with numbered placeholders (like "Connect every %1 days")
and one or more controls, and creates the GUI layout.

This helps a lot with translations, since now translators can work
with the entire sentence, instead of having to translate each label
around the text field separately.

This code is definitely not "ready" yet! Known issues:
 * All labels are left-aligned. Before this change, the left-most
   label was right-aligned to be closer to the text field (definitely
   looked better that way).
 * Tab order is incorrect if format string swaps controls (%2 before %1).
 * There are no unit tests for buildLayout.
 * There is no doxygen documentation for buildLayout or for UiFormatString.
 * buildLayout should be in its own file.